Db_customer modified and tested, address partly completed

+++ myshop-domain/src/main/java/com/db/entities/Db_Customer.java
@@ -28,7 +28,7 @@ public class Db_Customer {
 	@Column(name="password")
 	private String password;
 
-	public long getRid() {
+	public int getRid() {
 		return rid;
 	}
